backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
flask
flask-cors
python-dotenv
requests
google-generativeai
﻿blinker==1.9.0
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
Werkzeug==3.1.6
